refactor(web): extract SpatialEditor measurements and native listeners into hooks
PR 1 of the SpatialEditor decomposition (3957 -> 3789 lines), the two
extractions with no command-seam coupling:

- use-editor-measurements: rootSize/shellWidth ResizeObserver state,
  inspectorIsSheet breakpoint, viewportCenterScreen/containerSizeOf.
  MINIMAP_MIN_ROOT_WIDTH_PX moves here and is imported back.
- use-native-canvas-listeners: the { passive: false } wheel listener
  (via a latest-closure ref), touchstart/touchmove native-gesture
  refusal with the [data-editor-overlay] exemption, the document
  capture-phase gestureTrace recorder, and the unmount teardown
  (long-press timer + best-effort releasePointerCapture with the root
  captured at mount).

editor-state-surface.test.ts scans the editor PLUS the hooks its state
moved to (import.meta.glob list), so a moved useState stays ledgered;
mutation-checked by dropping the rootSize entry (fails naming it).

Behavior-preserving: spatial-editor jsdom 42 files / 468 tests and the
7 targeted browser files (66 tests) covering wheel zoom, minimap,
touch pan/zoom, lost pointer capture, inspector sheet/dock pass
un-weakened.
